Últimos cambios
- Campo bloqueado al editar usuario --> Listo
- ¿Se puede poner un modal de "está seguro de eliminar", por si la cagan en picarle sin querer a eliminar en lugar de editar XD --> Listo
- Quitar sub menu y dejar solo como "Inventario" --> Listo
- No jala botón Guardar, en editar productos :0 --> Listo
- Poner en editar, el código del proveedor, para los productos --> Listo
- Igual en eliminar productos, poner un botón está seguro de eliminar por si la cagan --> Listo
- En productos me carga una parte de la lista que está vacia :0 --> Listo
- Al registrar mercancia, así como al eliminarla, igual preguntar si está seguro de que los datos son correctos --> Listo
- En las partidas individuales, quitar el costo x kilo, el que debe tomar es el que resume los totales --> Listo
- Falta de grupo pecuario que pida el producto que le corresponde /../ --> Medio Listo

+++ src/app/inventario/pages/products/products.component.ts
@@ -46,38 +46,52 @@ export class ProductsComponent implements OnInit {
   openModalUpdate(id_product: number) {
     this.modalRefUpdate = this.modalService.open(ProductUpdateComponent, { data: { id_product } });
     this.modalRefUpdate.onClose.subscribe((msg: any) => {
+      localStorage.removeItem('id_producto');
       this.loadProducts();
     });
   }
 
   deleteProduct(id_product: number) {
-    this.productsService.deleteProduct(id_product).subscribe( resp => {
-      if ( resp.status === true ) {
-        Swal.fire({
-          title: 'Éxito',
-          text: resp.message,
-          icon: 'success',
-          confirmButtonText: 'Aceptar',
-          confirmButtonColor: '#0f1765',
-          customClass: {
-            container: 'my-swal'
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "¡No podrás revertir esto!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#0f1765',
+      cancelButtonColor: '#d33',
+      confirmButtonText: '¡Sí, elimínalo!',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productsService.deleteProduct(id_product).subscribe( resp => {
+          if ( resp.status === true ) {
+            Swal.fire({
+              title: 'Éxito',
+              text: resp.message,
+              icon: 'success',
+              confirmButtonText: 'Aceptar',
+              confirmButtonColor: '#0f1765',
+              customClass: {
+                container: 'my-swal'
+              }
+            }).then( () => {
+              this.loadProducts();
+            });
+          } else {
+            Swal.fire({
+              title: 'Error',
+              text: resp.message,
+              icon: 'error',
+              confirmButtonText: 'Aceptar',
+              confirmButtonColor: '#0f1765',
+              customClass: {
+                container: 'my-swal'
+              }
+            })
           }
-        }).then( () => {
-          this.loadProducts();
-        });
-      } else {
-        Swal.fire({
-          title: 'Error',
-          text: resp.message,
-          icon: 'error',
-          confirmButtonText: 'Aceptar',
-          confirmButtonColor: '#0f1765',
-          customClass: {
-            container: 'my-swal'
-          }
-        })
-      }
 
+        });
+      }
     });
   }
